data_platform/src/bdd-adf-pipelines/requirements.txt
azure-identity==1.5.0
azure-mgmt-datafactory==1.0.0
azure-mgmt-resource==16.0.0
azure-storage-file-datalake==12.3.0
behave==1.2.6
